refactor: begin major code and structure cleanup (WIP)
Started reorganizing Singularity library:
- Removed obsolete trait headers
- Added initial Core and Types structure
- Work in progress, not complete
