+++ Singularity/Core/Enums.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+namespace Sglty::Core {
+
+enum class Mode { Dense, Sparse };
+
+enum class Order { RowMajor, ColMajor, Undefined };
+
+}  // namespace Sglty::Core
+
+// Singularity/Core/Enums.hpp
