+++ Singularity/Expr/Evaluate.hpp
@@ -1,24 +1,21 @@
-#ifndef SINGULARITYLIB_EXPR_EVALUATOR_HPP
-#define SINGULARITYLIB_EXPR_EVALUATOR_HPP
+#pragma once
 
 #include "../Core/Checks.hpp"
-#include "../Matrix.hpp"
 #include "../Traits/Expr.hpp"
+#include "../Types/Matrix.hpp"
 #include "Base.hpp"
 
-namespace Sglty {
-
-namespace Expr {
+namespace Sglty::Expr {
 
 template <typename _expr>
 constexpr auto Evaluate(const _expr& _e) -> std::enable_if_t<
     is_expression_v<_expr>,
-    Matrix<typename Traits::Extract<
+    Sglty::Types::Matrix<typename Sglty::Traits::Extract<
         _expr>::template core_rebind<_expr::rows, _expr::cols>>> {
   using core_t =
-      typename Traits::Extract<_expr>::template core_rebind<_expr::rows,
-                                                            _expr::cols>;
-  using ret_type  = Matrix<core_t>;
+      typename Sglty::Traits::Extract<_expr>::template core_rebind<_expr::rows,
+                                                                   _expr::cols>;
+  using ret_type  = Sglty::Types::Matrix<core_t>;
   using size_type = typename ret_type::size_type;
 
   ret_type result;
@@ -33,12 +30,12 @@ constexpr auto Evaluate(const _expr& _e) -> std::enable_if_t<
 
 template <typename _core_ret, typename _expr>
 constexpr auto Evaluate(const _expr& _e) -> std::enable_if_t<
-    Core::is_valid_implementation_v<_core_ret> && is_expression_v<_expr>,
-    Matrix<
+    Core::is_valid_implementation_v<_core_ret> && Sglty::is_expression_v<_expr>,
+    Sglty::Types::Matrix<
         typename _core_ret::template core_rebind<_expr::rows, _expr::cols>>> {
   using core_t =
       typename _core_ret::template core_rebind<_expr::rows, _expr::cols>;
-  using ret_type  = Matrix<core_t>;
+  using ret_type  = Sglty::Types::Matrix<core_t>;
   using size_type = typename ret_type::size_type;
 
   ret_type result;
@@ -50,10 +47,6 @@ constexpr auto Evaluate(const _expr& _e) -> std::enable_if_t<
   return result;
 }
 
-}  // namespace Expr
-
-}  // namespace Sglty
-
-#endif  // SINGULARITYLIB_EXPR_EVALUATOR_HPP
+}  // namespace Sglty::Expr
 
 // Singularity/Expr/Evaluator.hpp
