--- Singularity/Lib/Algebra.hpp
@@ -1,8 +1,0 @@
-#ifndef SINGULARITY_ALGEBRA_HPP
-#define SINGULARITY_ALGEBRA_HPP
-
-#include "Algebra/Trp.hpp"
-
-#endif  // SINGULARITY_ALGEBRA_HPP
-
-// Singularity/Algebra.hpp
