--- Singularity/Lib/Arithmetic.hpp
@@ -1,10 +1,0 @@
-#ifndef SINGULARITY_ARITHMETIC_HPP
-#define SINGULARITY_ARITHMETIC_HPP
-
-#include "Arithmetic/Add.hpp"
-#include "Arithmetic/Mul.hpp"
-#include "Arithmetic/Sub.hpp"
-
-#endif  // SINGULARITY_ARITHMETIC_HPP
-
-// Singularity/Arithmetic.hpp
