--- Singularity/Lib/Core.hpp
@@ -1,11 +1,0 @@
-#ifndef SINGULARITY_CORE_HPP
-#define SINGULARITY_CORE_HPP
-
-#include "Core/Checks.hpp"
-#include "Core/Dense.hpp"
-#include "Core/Enums.hpp"
-#include "Core/Sparse.hpp"
-
-#endif  // SINGULARITY_CORE_HPP
-
-// Singularity/Core.hpp
