--- Singularity/Lib/Core/Enums.hpp
@@ -1,18 +1,0 @@
-#ifndef SINGULARITY_CORE_ENUMS_HPP
-#define SINGULARITY_CORE_ENUMS_HPP
-
-namespace Sglty {
-
-namespace Core {
-
-enum class Mode { Dense, Sparse };
-
-enum class Order { RowMajor, ColMajor, Undefined };
-
-}  // namespace Core
-
-}  // namespace Sglty
-
-#endif  // SINGULARITY_CORE_ENUMS_HPP
-
-// Singularity/Core/Enums.hpp
