--- Singularity/Lib/Expr.hpp
@@ -1,9 +1,0 @@
-#ifndef SINGULARITY_EXPR_HPP
-#define SINGULARITY_EXPR_HPP
-
-#include "Expr/Base.hpp"
-#include "Expr/Evaluate.hpp"
-
-#endif  // SINGULARITY_EXPR_HPP
-
-// Singularity/Expr.hpp
