--- Singularity/Lib/Traits.hpp
@@ -1,12 +1,0 @@
-#ifndef SINGULARITY_TRAITS_HPP
-#define SINGULARITY_TRAITS_HPP
-
-#include "Traits/Core.hpp"
-#include "Traits/Expr.hpp"
-#include "Traits/Matrix.hpp"
-#include "Traits/Size.hpp"
-#include "Traits/Type.hpp"
-
-#endif  // SINGULARITY_TRAITS_HPP
-
-// Singularity/Traits.hpp
