+++ Singularity/Traits/Matrix.hpp
@@ -1,5 +1,4 @@
-#ifndef SINGULARITY_TRAITS_MATRIX_HPP
-#define SINGULARITY_TRAITS_MATRIX_HPP
+#pragma once
 
 #include <type_traits>
 
@@ -20,7 +19,5 @@ template <typename _expr>
 constexpr bool is_matrix_v = IsMatrix<_expr>::value;
 
 }  // namespace Sglty
-
-#endif  // SINGULARITY_TRAITS_MATRIX_HPP
 
 // Singularity/Traits/Matrix.hpp
